- refactoring to separate common and postgres-specific parts of API
